requirements.txt
canopy-sdk==0.9.0
chardet==5.2.0
pinecone-client==3.2.2
pinecone-text==0.8.0
python-multipart==0.0.6
python-dotenv==1.0.1
fastapi==0.110.1
